LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-exit-fullscreen-nested-shadow-dom-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected Element node <div id="outer">
      <slot></slot>
    </div> but got null
fullscreen

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected Element node <div id="outer">
      <slot></slot>
    </div> but got null

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected Element node <div id="outer">
      <slot></slot>
    </div> but got null

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out
